move checks to their own file
