+++ sklearn/tests/test_docstring_parameters.py
@@ -23,6 +23,10 @@ from sklearn.experimental import (
 from sklearn.linear_model import LogisticRegression
 from sklearn.preprocessing import FunctionTransformer
 from sklearn.utils import all_estimators
+from sklearn.utils._test_common._common import (
+    _enforce_estimator_tags_X,
+    _enforce_estimator_tags_y,
+)
 from sklearn.utils._test_common.instance_generator import _construct_instance
 from sklearn.utils._testing import (
     _get_func_name,
@@ -31,10 +35,6 @@ from sklearn.utils._testing import (
     ignore_warnings,
 )
 from sklearn.utils.deprecation import _is_deprecated
-from sklearn.utils.estimator_checks import (
-    _enforce_estimator_tags_X,
-    _enforce_estimator_tags_y,
-)
 
 # walk_packages() ignores DeprecationWarnings, now we need to ignore
 # FutureWarnings
